chore: move data-service module to the right folder
