+++ apps/backend/src/modules/app.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { DataServicesModule } from '../infra/data-services.module';
+import { DataServicesModule } from './data-services.module';
 import { AuthModule } from './auth.module';
 import { UserModule } from './user.module';
 import { PetModule } from './pet.module';
